Moved energy/flux/etc functions into energy.jl in order to modularise the code. Passes tests
