+++ agni.jl
@@ -17,6 +17,7 @@ using TOML
 # Include local jl files
 push!(LOAD_PATH, joinpath(ROOT_DIR,"src"))
 import atmosphere
+import energy
 import setpt
 import plotting 
 import phys
@@ -320,7 +321,7 @@ function main()::Bool
     end 
 
     # Solver variables 
-    incl_convect::Bool = !isempty(conv_type)
+    incl_convect::Bool= !isempty(conv_type)
     use_mlt::Bool     = (conv_type == "mlt")
     modplot::Int      = 0
     incl_conduct::Bool = false
@@ -347,17 +348,17 @@ function main()::Bool
         # No solve - just calc fluxes at the end
         if sol == "none"
             fill!(atmos.flux_tot, 0.0)
-            atmosphere.radtrans!(atmos, true, calc_cf=true)
-            atmosphere.radtrans!(atmos, false)
+            energy.radtrans!(atmos, true, calc_cf=true)
+            energy.radtrans!(atmos, false)
             if use_mlt 
-                atmosphere.mlt_dry!(atmos)
+                energy.mlt!(atmos)
             end 
             if incl_sens 
-                atmosphere.sensible!(atmos)
+                energy.sensible!(atmos)
             end 
-            atmosphere.condense_relax!(atmos, condensates)
+            energy.condense_relax!(atmos, condensates)
             if incl_conduct
-                atmosphere.conduct!(atmos)
+                energy.conduct!(atmos)
             end
             atmos.flux_tot = atmos.flux_cdry + atmos.flux_n + atmos.flux_cdct + atmos.flux_p
             atmos.flux_tot[end] += atmos.flux_sens
@@ -371,7 +372,7 @@ function main()::Bool
             end
             solver_success = solver_tstep.solve_energy!(atmos, sol_type=sol_type, use_physical_dt=false,
                                 modplot=modplot, modprop=5, verbose=true,  sens_heat=incl_sens, chem_type=chem_type,
-                                incl_convect=incl_convect, condensates=condensates, conduct=incl_conduct,
+                                convect=incl_convect, condensates=condensates, conduct=incl_conduct,
                                 accel=stabilise, step_rtol=1.0e-4, step_atol=1.0e-2, dt_max=1000.0,
                                 conv_atol=conv_atol, conv_rtol=conv_rtol, save_frames=plt_ani,
                                 max_steps=max_steps, min_steps=100, use_mlt=use_mlt)
@@ -385,7 +386,7 @@ function main()::Bool
             method = findfirst(==(sol), method_map)
             solver_success = solver_nlsol.solve_energy!(atmos, sol_type=sol_type, 
                                 conduct=incl_conduct,  chem_type=chem_type,
-                                incl_convect=incl_convect, condensates=condensates, sens_heat=incl_sens,
+                                convect=incl_convect, condensates=condensates, sens_heat=incl_sens,
                                 max_steps=max_steps, conv_atol=conv_atol, conv_rtol=conv_rtol, method=1,
                                 stabilise_mlt=stabilise,modplot=modplot,save_frames=plt_ani)
             return_success = return_success && solver_success
