+++ demo_steamrun.jl
@@ -76,7 +76,7 @@ for i in 1:run_len
     setpt.condensing!(atmos, "H2O")
 
     # Calculate LW fluxes 
-    atmosphere.radtrans!(atmos, true)
+    energy.radtrans!(atmos, true)
 
     olr = atmos.flux_u_lw[1]
     @info @sprintf("Tsurf = %4.1f K  ,  OLR = %5.1f W m-2 ",atmos.tmpl[end],olr)
